bfs: Remove unnecessary resched_suitable_idle() in schedule().
When prev task is not the idle task and needs other cpu, the next
task will be returned from earliest_deadline_task() and will be
not the same task as prev task because the affinity.
resched_suitable_idle() will be called after prev != next check,
and resched_suitable_idle() call after needs_other_cpu() is not
necessary.