bfs: [Fix] Add additional checking in sched_submit_work()
When sched_submit_works() is introduced in mainline 3.1, it moves the
blk_schedule_flush_plug(tsk) call outside from schedule(), but relaxing
the checking when not calling it. The code change is ok for mainline
CFS but it's not for BFS somehow.

Adding back those checking is the current solution for BFS.